refactor:moved initiative code from leader feed screen to initiative profile screen
